requirements.txt
inflect~=7.0.0
openai~=0.27.9
pyyaml~=6.0.1
python-dotenv~=1.0.0
transformers~=4.32.0
argparse~=1.4.0
tqdm~=4.66.1
xmltodict~=0.13.0
google-generativeai~=0.1.0
torch~=2.0.1